requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tensorflow
python-chess
supabase
python-dotenv
sendgrid
bcrypt